Move client-safe ETF factor helpers to own module so fs-backed flag loader stays server-only (fix client bundle build)
